% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/converg.R
\name{converg}
\alias{converg}
\title{Graph of Running Statistics for Convergence Diagnostics}
\usage{
converg(
  x,
  node = length(x),
  margin = c("var", "unc"),
  nvariates = 1,
  iter = 1,
  probs = c(0.025, 0.975),
  lim = c(0.025, 0.975),
  griddim = NULL,
  log = FALSE
)
}
\arguments{
\item{x}{a \code{\link{mcnode}}, \code{\link{mc}} or \code{\link{mccut}} object.}

\item{node}{the node to be considered in a \samp{mc} or \samp{mccut} object, either as
order number or name. By default: the last node. Cannot be of type \samp{"0"}.}

\item{margin}{the margin used to plot the graph. Used only if the node is a \samp{"VU"} node.}

\item{nvariates}{the variate to be considered. Used only for multivariate nodes.}

\item{iter}{if \samp{margin=="var"} and the node is \samp{"VU"}, specifies which iteration
in the uncertainty dimension to use.}

\item{probs}{the quantiles to be provided in the variability dimension.}

\item{lim}{the quantiles to be used in the uncertainty dimension.}

\item{griddim}{a vector of two integers for the grid size. If \samp{NULL}, calculated automatically.}

\item{log}{if \samp{TRUE}, the data will be log-transformed.}
}
\value{
\samp{invisible(NULL)} (called for its side effect of drawing a plot).
}
\description{
Provides basic graphs to evaluate the convergence of a node of a \code{\link{mc}} or
a \code{\link{mccut}} object in the variability or in the uncertainty dimension.
}
\details{
If the node is of type \samp{"V"}, the running mean, median and \samp{probs} quantiles
according to the variability dimension will be provided. If the node is of type \samp{"VU"}
and \samp{margin="var"}, this graph will be provided on one simulation in the uncertainty
dimension (chosen by \samp{iter}).

If the node is of type \samp{"U"}, the running mean, median and \samp{lim} quantiles
according to the uncertainty dimension will be provided.

If the node is of type \samp{"VU"} (with \samp{margin="unc"} or from a \samp{mccut} object),
one graph is provided for each of the mean, median and \samp{probs} quantiles calculated in
the variability dimension.
}
\note{
This function may be used on a \samp{mccut} object only if a \samp{summary.mc}
function was used in the third block of the \code{\link{evalmccut}} call. The values
used as \samp{probs} in \samp{converg} should have been used in that \samp{summary.mc}.
}
\examples{
data(total)
converg(xVU, margin="var")
converg(xVU, margin="unc")
}
\keyword{hplot}
