% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cornode.R
\name{cornode}
\alias{cornode}
\title{Build a Rank Correlation Using the Iman and Conover Method}
\usage{
cornode(..., target, outrank = FALSE, result = FALSE, seed = NULL)
}
\arguments{
\item{...}{a matrix (each of its \samp{n} columns but the first one will be reordered) or
\samp{n} \samp{mcnode} objects (each element but the first will be reordered). Arguments
should be named.}

\item{target}{a scalar (only if \samp{n=2}) or a \samp{(n x n)} matrix of correlations.}

\item{outrank}{should the order be returned?}

\item{result}{should the correlation eventually obtained be printed?}

\item{seed}{the random seed for building the correlation. If \samp{NULL} the seed is unchanged.}
}
\value{
If \samp{outrank=FALSE}: the matrix or a list of rearranged \samp{mcnode}s.
  If \samp{outrank=TRUE}: the order to be used to rearrange to build the desired correlation.
}
\description{
Builds a rank correlation structure between columns of a matrix or between \samp{mcnode}
objects using the Iman and Conover (1982) method.
}
\details{
The function accepts for \samp{data}:
\itemize{
  \item some \samp{"V" mcnode} objects separated by a comma;
  \item some \samp{"U" mcnode} objects separated by a comma;
  \item some \samp{"VU" mcnode} objects, building the structure column by column;
  \item one \samp{"V" mcnode} as first element and some \samp{"VU" mcnode} objects.
}

\samp{target} should be a scalar (two columns only) or a real symmetric positive-definite
square matrix. Only the upper triangular part is used (see \code{\link{chol}}).

The final correlation structure should be checked, as it is not always possible to achieve
the target exactly. The order of values within each \samp{mcnode} (except the first) will
be changed by this function. The \samp{outrank} option may help to rebuild prior links.
}
\examples{
x1 <- rnorm(1000)
x2 <- rnorm(1000)
x3 <- rnorm(1000)
mat <- cbind(x1, x2, x3)
(corr <- matrix(c(1,0.5,0.2, 0.5,1,0.2, 0.2,0.2,1), ncol=3))
cor(mat, method="spearman")
matc <- cornode(mat, target=corr, result=TRUE)
all(matc[,1] == mat[,1])
}
\references{
Iman, R. L., & Conover, W. J. (1982). A distribution-free approach to inducing
  rank correlation among input variables. \emph{Communication in Statistics - Simulation and
  Computation}, 11(3), 311-334.
}
\keyword{multivariate}
