% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.mc.R
\name{dimmcnode}
\alias{dimmcnode}
\alias{dimmc}
\alias{typemcnode}
\alias{is.mc}
\alias{is.mcnode}
\title{Dimension of mcnode and mc Objects}
\usage{
dimmcnode(x)

dimmc(x)

typemcnode(x, index = FALSE)

is.mc(x)

is.mcnode(x)
}
\arguments{
\item{x}{a \samp{mcnode} or a \samp{mc} object.}

\item{index}{if \samp{TRUE}, give the index of the type rather than the type (for
\samp{typemcnode} only).}
}
\value{
\samp{dimmcnode}: a vector of three scalars (nsv, nsu, nvariates).
  \samp{dimmc}: a vector of three scalars (max nsv, max nsu, max variates).
  \samp{typemcnode}: \samp{"0"}, \samp{"V"}, \samp{"U"} or \samp{"VU"}, or the
  corresponding index if \samp{index=TRUE}. \samp{NULL} if none found.
  \samp{is.mc}: \samp{TRUE} or \samp{FALSE}.
  \samp{is.mcnode}: \samp{TRUE} or \samp{FALSE}.
}
\description{
Provides the dimension of a \samp{mcnode} or a \samp{mc} object: the number of simulations
in the variability dimension, in the uncertainty dimension, and the maximum number of
variates.
}
\note{
These functions do not test whether the object is correctly built. See
  \code{\link{is.mcnode}} and \code{\link{is.mc}}.
}
\examples{
data(total)
dimmcnode(xVUM2)
dimmc(total)
typemcnode(total$xVUM2)
is.mcnode(xVU)
is.mc(total)
}
\seealso{
\code{\link{mcnode}}, \code{\link{mc}}.
}
\keyword{utilities}
