% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rdirichlet.R
\name{dirichlet}
\alias{dirichlet}
\alias{ddirichlet}
\alias{rdirichlet}
\title{The Dirichlet Distribution}
\usage{
ddirichlet(x, alpha)

rdirichlet(n, alpha)
}
\arguments{
\item{x}{a vector containing a single deviate or a matrix containing one random deviate per row.}

\item{alpha}{a vector of shape parameters, or a matrix of shape parameters by rows.
Recycling (by row) is permitted.}

\item{n}{number of random vectors to generate. If \samp{length(n) > 1}, the length is
taken to be the number required.}
}
\value{
\samp{ddirichlet} gives the density. \samp{rdirichlet} returns a matrix with
\samp{n} rows, each containing a single Dirichlet random deviate.
}
\description{
Density function and random generation from the Dirichlet distribution.
}
\details{
The Dirichlet distribution is the multidimensional generalization of the beta distribution.
The original code was adapted to provide a kind of "vectorization" used in multivariate
\samp{mcnode} objects.
}
\note{
Code is adapted from \samp{MCMCpack}. It originates from Greg's Miscellaneous
Functions (gregmisc).
}
\examples{
dat <- c(1, 10, 100, 1000, 1000, 100, 10, 1)
(alpha <- matrix(dat, nrow = 4, byrow = TRUE))
round(x <- rdirichlet(4, alpha), 2)
ddirichlet(x, alpha)

## rdirichlet used with mcstoc
mcalpha <- mcdata(dat, type = "V", nsv = 4, nvariates = 2)
(x <- mcstoc(rdirichlet, type = "V", alpha = mcalpha, nsv = 4, nvariates = 2))
unclass(x)
}
\seealso{
\code{\link{Beta}}
}
\keyword{distribution}
