% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmultinomial.R
\name{dmultinomial}
\alias{dmultinomial}
\alias{rmultinomial}
\title{The Vectorized Multinomial Distribution}
\usage{
dmultinomial(x, size = NULL, prob, log = FALSE)

rmultinomial(n, size, prob)
}
\arguments{
\item{x}{vector or matrix of length (or ncol) K of integers in \samp{0:size}.}

\item{size}{a vector of integers, say N, specifying the total number of objects that are
put into K boxes in the typical multinomial experiment. For \samp{dmultinomial}, it
defaults to \samp{sum(x)}. The first element corresponds to the vector \samp{prob} or
the first row of \samp{prob}, ...}

\item{prob}{numeric non-negative vector of length K, or matrix of size \samp{(n x K)}
specifying the probability for the K classes; is internally normalized to sum 1.}

\item{log}{logical; if \samp{TRUE}, log probabilities are computed.}

\item{n}{number of random vectors to draw.}
}
\description{
Generate multinomially distributed random number vectors and compute multinomial
probabilities. These functions are the vectorized versions of \code{\link{rmultinom}}
and \code{\link{dmultinom}}. Recycling is permitted.
}
\examples{
x <- c(100, 200, 700)
x1 <- matrix(c(100,200,700, 200,100,700, 700,200,100), byrow=TRUE, ncol=3)
p <- c(1, 2, 7)
p1 <- matrix(c(1,2,7, 2,1,7, 7,2,1), byrow=TRUE, ncol=3)
dmultinomial(x1, prob=p)
## is equivalent to
c(dmultinom(x1[1,], prob=p),
  dmultinom(x1[2,], prob=p),
  dmultinom(x1[3,], prob=p))

prob <- c(1, 2, 7)
rmultinomial(4, 1000, prob)
rmultinomial(4, c(10, 100, 1000, 10000), prob)
}
\keyword{distribution}
