% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empiricalD.R
\name{empiricalD}
\alias{empiricalD}
\alias{dempiricalD}
\alias{pempiricalD}
\alias{qempiricalD}
\alias{rempiricalD}
\title{The Discrete Empirical Distribution}
\usage{
dempiricalD(x, values, prob = NULL, log = FALSE)

pempiricalD(q, values, prob = NULL, lower.tail = TRUE, log.p = FALSE)

qempiricalD(p, values, prob = NULL, lower.tail = TRUE, log.p = FALSE)

rempiricalD(n, values, prob = NULL)
}
\arguments{
\item{x, q}{vector of quantiles.}

\item{values}{vector or matrix of numerical values. See details.}

\item{prob}{optional vector or matrix of counts or probabilities. See details.}

\item{log, log.p}{logical; if \samp{TRUE}, probabilities \samp{p} are given as \samp{log(p)}.}

\item{lower.tail}{logical; if \samp{TRUE} (default), probabilities are \samp{P[X <= x]},
otherwise, \samp{P[X > x]}.}

\item{p}{vector of probabilities.}

\item{n}{number of random values. If \samp{length(n) > 1}, the length is taken to be
the number required.}
}
\value{
\samp{dempiricalD} gives the density, \samp{pempiricalD} gives the distribution
function, \samp{qempiricalD} gives the quantile function and \samp{rempiricalD} generates
random deviates.
}
\description{
Density, distribution function, quantile function and random generation for a discrete
empirical distribution. This function is vectorized to accept different sets of
\samp{values} or \samp{prob}.
}
\details{
If \samp{prob} is missing, the discrete distribution is obtained directly from the
vector of \samp{values}, otherwise \samp{prob} is used to weight the values.
\samp{prob} is normalized before use. Thus, \samp{prob} may be the count of each
\samp{values}. \samp{prob} values should be non negative and their sum should not be 0.

\samp{values} and/or \samp{prob} may vary: in that case, \samp{values} and/or \samp{prob}
should be sent as matrices, the first row being used for the first element of \samp{x},
\samp{q}, \samp{p} or the first random value, the second row for the second element, ...
Recycling is permitted if the number of rows of \samp{prob} and \samp{values} are equal
or if the number of rows of \samp{prob} and/or \samp{values} is one.

\samp{rempiricalD(n, values, prob)} with \samp{values} and \samp{prob} as vectors is
equivalent to \samp{sample(x=values, size=n, replace=TRUE, prob=prob)}.
}
\note{
In the future, the functions should be written for non numerical values.
}
\examples{
dempiricalD(1:6, 2:6, prob=c(10, 10, 70, 0, 10))
pempiricalD(1:6, 2:6, prob=c(10, 10, 70, 0, 10))
qempiricalD(seq(0, 1, 0.1), 2:6, prob=c(10, 10, 70, 0, 10))
table(rempiricalD(10000, 2:6, prob=c(10, 10, 70, 0, 10)))

## Varying values
(values <- matrix(1:10, ncol=5))
dempiricalD(c(1, 1), values)
}
\seealso{
\code{\link{sample}}, \code{\link{empiricalC}}.
}
\keyword{distribution}
