% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evalmcmod.R
\name{evalmcmod}
\alias{evalmcmod}
\title{Evaluates a Monte Carlo Model}
\usage{
evalmcmod(expr, nsv = ndvar(), nsu = ndunc(), seed = NULL)
}
\arguments{
\item{expr}{a model of class \code{\link{mcmodel}} or a valid expression.}

\item{nsv}{the number of simulations in the variability dimension.}

\item{nsu}{the number of simulations in the uncertainty dimension.}

\item{seed}{the random seed. If \samp{NULL} the seed is unchanged.}
}
\value{
The result of the evaluation — should be a \samp{mc} object.
}
\description{
Evaluates a \code{\link{mcmodel}} object (or a valid expression) using a specified number
of simulations and with (or without) a specified seed.
}
\details{
The model is evaluated. Intermediate variables used to build the \samp{mc} object are not
stored.
}
\note{
The seed is set at the beginning of the evaluation. Thus, the complete similarity
of two evaluations with the same seed is not certain, depending on the structure of the model.
}
\examples{
data(ec)
ec$modEC1
evalmcmod(ec$modEC1, nsv=100, nsu=100, seed=666)
}
\seealso{
\code{\link{mcmodel}}, \code{\link{evalmccut}} to evaluate high-dimension models.
}
\keyword{methods}
