% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractvar.R
\name{extractvar}
\alias{extractvar}
\alias{addvar}
\title{Utilities for Multivariate Nodes}
\usage{
extractvar(x, which = 1)

addvar(...)
}
\arguments{
\item{x}{a multivariate \samp{mcnode}.}

\item{which}{a vector specifying which variate(s) to extract.}

\item{...}{\samp{mcnode}s to be combined in a multivariate \samp{mcnode}.
These \samp{mcnode}s should be of the same type and dimension.}
}
\value{
The new \samp{mcnode}.
}
\description{
\samp{extractvar} extracts one or more variates from a multivariate \samp{mcnode}.
\samp{addvar} combines consistent \samp{mcnode}s into a single multivariate \samp{mcnode}.
}
\details{
The \samp{outm} attribute of the output of \samp{addvar} is taken from the first element.
}
\examples{
x <- mcdata(0:3, "0", nvariates = 4)
y <- extractvar(x, c(1, 3))
y
addvar(x, y)
}
\seealso{
\code{\link{mcnode}} for \samp{mcnode} objects.
}
\keyword{methods}
