% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist.mc.R
\name{hist.mc}
\alias{hist.mc}
\alias{hist.mcnode}
\title{Histogram of a Monte Carlo Simulation}
\usage{
\method{hist}{mc}(x, griddim = NULL, xlab = names(x), ylab = "Frequency", main = "", ...)

\method{hist}{mcnode}(x, ...)
}
\arguments{
\item{x}{an \samp{mcnode} or an \samp{mc} object.}

\item{griddim}{a vector of two integers, indicating the size of the grid of plots.
If \samp{NULL}, the grid is calculated to produce a "nice" graph.}

\item{xlab}{a vector of labels for the x-axis. May be recycled.}

\item{ylab}{a vector of labels for the y-axis. May be recycled.}

\item{main}{a vector of main titles of histograms. May be recycled.}

\item{...}{other arguments to be passed to all calls of \samp{hist}.}
}
\value{
\samp{invisible()} (called for side effects).
}
\description{
Shows a histogram of a \samp{mcnode} or a \samp{mc} object.
}
\note{
For two-dimensional \samp{mc} objects, the histogram is based on all data
(variability and uncertainty) pooled together.
}
\examples{
data(total)
hist(xVUM3)
hist(total)
}
\seealso{
\code{\link{gghist}} for a ggplot2 version.
}
\keyword{hplot}
