% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lhs.R
\name{lhs}
\alias{lhs}
\title{Random Latin Hypercube Sampling}
\usage{
lhs(distr = "runif", nsv = ndvar(), nsu = ndunc(), nvariates = 1, ...)
}
\arguments{
\item{distr}{the function for generating random samples, or its name as a character string.
If \samp{distr} is \samp{"rdist"}, the quantile function \samp{"qdist"} must exist with
argument \samp{p} as a vector of probabilities.}

\item{nsv}{the number of rows of the final matrix.}

\item{nsu}{the number of columns of the final matrix.}

\item{nvariates}{the number of variates.}

\item{...}{all arguments to be passed to \samp{distr} except the size of the sample.}
}
\value{
A \samp{nsv x nsu} matrix of random variates.
}
\description{
Creates a Latin Hypercube Sample (LHS) from the specified distribution.
}
\note{
The resulting LHS is a latin hypersquare sampling: the LHS structure is provided
only in the first 2 dimensions.
It is not possible to use truncated distributions directly with \code{\link{rtrunc}}.
Use \code{\link{mcstoc}} with \samp{lhs=TRUE} and \samp{rtrunc=TRUE} for that purpose.
The \dots arguments will be recycled.

Adapted from a code by Rob Carnell (package \pkg{lhs}).
}
\examples{
ceiling(lhs(runif, nsu=10, nsv=10) * 10)
}
\seealso{
\code{\link{mcstoc}}
}
\keyword{design}
