% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mccontrol.R
\name{mc.control}
\alias{mc.control}
\alias{ndvar}
\alias{ndunc}
\title{Sets or Gets the Default Number of Simulations}
\usage{
ndvar(n)

ndunc(n)
}
\arguments{
\item{n}{number of simulations.}
}
\value{
The current value, AFTER modification if \samp{n} is present (unlike \code{options}).
}
\description{
Sets or retrieves the default number of simulations in the variability and
uncertainty dimensions.
}
\details{
\samp{ndvar()} gets and \samp{ndvar(n)} sets the default number of simulations in
the 1D simulations or the number of simulations in the variability dimension in
2D simulations.

\samp{ndunc()} gets and \samp{ndunc(n)} sets the number of simulations in the
uncertainty dimension in 2D simulations.

\samp{n} is rounded to its ceiling value.

The default values when loaded are 1001 for \samp{ndvar} and 101 for \samp{ndunc}.
}
\examples{
(oldvar <- ndvar())
(oldunc <- ndunc())
mcstoc(runif, type = "VU")
ndvar(12)
ndunc(21)
mcstoc(runif, type = "VU")
ndvar(oldvar)
ndunc(oldunc)
}
\seealso{
\code{\link{mcstoc}}, \code{\link{mcdata}}
}
\keyword{misc}
