% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc.R
\name{mc}
\alias{mc}
\title{Monte Carlo Object}
\usage{
mc(..., name = NULL, devname = FALSE)
}
\arguments{
\item{...}{\samp{mcnode} and/or \samp{mc} object(s) to be gathered in a \samp{mc} object,
separated by a comma.}

\item{name}{vector of characters of the same length as the final \samp{mc} object.
If \samp{NULL}, names are taken from the names of the elements.}

\item{devname}{develop the name from the name of the \samp{mc} objects, if any.}
}
\value{
An object of class \samp{mc}.
}
\description{
Creates \samp{mc} objects from \code{\link{mcnode}} or \samp{mc} objects.
}
\details{
A \samp{mc} object is a list of \code{\link{mcnode}} objects.
\samp{mcnode} objects must be of coherent dimensions.

If one of the arguments is a \samp{mc} object, the names of its elements are used.
\samp{devname = TRUE} develops the name using the name of the \samp{mc} object as prefix.
Finally, names are made unique.
}
\examples{
x <- mcstoc(runif)
y <- mcdata(3, type = "0")
z <- x * y
(m <- mc(x, y, z, name = c('n1', 'n2', 'n3')))
mc(m, x, devname = TRUE)
}
\seealso{
\code{\link{mcnode}}, the basic element of a \samp{mc} object.
  To evaluate \samp{mc} objects: \code{\link{mcmodel}}, \code{\link{evalmcmod}},
  \code{\link{evalmccut}}.
  To study \samp{mc} objects: \code{\link{print.mc}}, \code{\link{summary.mc}},
  \code{\link{plot.mc}}, \code{\link{converg}}, \code{\link{hist.mc}},
  \code{\link{tornado}}, \code{\link{tornadounc}}.
}
\keyword{methods}
