% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mc2d-package.R
\docType{package}
\name{mc2d-package}
\alias{mc2d}
\alias{mc2d-package}
\title{mc2d: Tools for Two-Dimensional Monte-Carlo Simulations}
\description{
A complete framework to build and study Two-Dimensional Monte-Carlo simulations, aka Second-Order Monte-Carlo simulations. Also includes various distributions (pert, triangular, Bernoulli, empirical discrete and continuous).
}
\author{
\strong{Maintainer}: Regis Pouillot \email{rpouillot@yahoo.fr}

Authors:
\itemize{
  \item Regis Pouillot \email{rpouillot@yahoo.fr}
}

Other contributors:
\itemize{
  \item Marie-Laure Delignette-Muller \email{ml.delignette@vetagro-sup.fr} [contributor]
  \item Jean-Baptiste Denis \email{jbdenis@jouy.inra.fr} [contributor]
  \item Yu Chen \email{chen0099yu@gmail.com} [contributor]
  \item Arie Havelaar \email{ariehavelaar@ufl.edu} [contributor]
}

}
