% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcapply.R
\name{mcapply}
\alias{mcapply}
\title{Apply Functions Over mc or mcnode Objects}
\usage{
mcapply(x, margin = c("all", "var", "unc", "variates"), fun, ...)
}
\arguments{
\item{x}{a \samp{mc} or a \samp{mcnode} object.}

\item{margin}{the dimension on which to apply the function. May be \samp{"all"} (default)
to apply the function on all values, \samp{"var"} to apply on the variability dimension,
\samp{"unc"} to apply on the uncertainty dimension, or \samp{"variates"} to apply on
the variates. Note: do not use \samp{"var"} for \samp{"variates"}.}

\item{fun}{the function to be applied. When applied to a vector of length \samp{n},
\samp{fun} should return a vector of length \samp{n} or \samp{1}.}

\item{...}{optional arguments to \samp{fun}.}
}
\value{
If \samp{fun} returns a vector of length \samp{n} or if \samp{margin="all"},
  the returned \samp{mcnode}s have the same type and dimension as \samp{x}.
  Otherwise, the type of \samp{mcnode} is changed accordingly.
}
\description{
Apply a function on all values or over a given dimension of an \samp{mcnode} object.
May be used for all \samp{mcnode} objects of an \samp{mc} object.
}
\examples{
data(total)
xVUM
mcapply(xVUM, "unc", sum)
mcapply(xVUM, "var", sum)
mcapply(xVUM, "all", sum)
mcapply(xVUM, "variates", sum)
mcapply(total, "all", exp)
}
\seealso{
\code{\link{apply}}, \code{\link{mc}}, \code{\link{mcnode}}.
}
\keyword{misc}
