% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcmodel.R
\name{mcmodel}
\alias{mcmodel}
\title{Specify a Monte Carlo Model}
\usage{
mcmodel(x, is.expr = FALSE)
}
\arguments{
\item{x}{an \R call or an expression.}

\item{is.expr}{\samp{FALSE} to send a call, \samp{TRUE} to send an expression.}
}
\value{
an \R expression of class \samp{mcmodel}.
}
\description{
Specifies a \samp{mcmodel} without evaluating it, for subsequent evaluation using
\code{\link{evalmcmod}}.
}
\details{
The model should be put between \samp{\{} and \samp{\}}. The last line should be of the
form \samp{mc(...)}. Any reference to the number of simulations in the variability
dimension should use \samp{ndvar()} or (preferred) \samp{nsv}. Any reference to the
number of simulations in the uncertainty dimension should use \samp{ndunc()} or (preferred)
\samp{nsu}.
}
\examples{
modEC1 <- mcmodel({
  conc <- mcdata(10, "0")
  cook <- mcstoc(rempiricalD, values=c(0, 1/5, 1/50), prob=c(0.027, 0.373, 0.600))
  serving <- mcstoc(rgamma, shape=3.93, rate=0.0806)
  expo <- conc * cook * serving
  dose <- mcstoc(rpois, lambda=expo)
  risk <- 1 - (1 - 0.001)^dose
  mc(conc, cook, serving, expo, dose, risk)
})
evalmcmod(modEC1, nsv=100, nsu=100)
}
\seealso{
\code{\link{expression}}, \code{\link{evalmcmod}} to evaluate the model,
  \code{\link{mcmodelcut}} to evaluate in a loop.
}
\keyword{methods}
