% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcdata.R, R/mcdatanocontrol.R
\name{mcnode}
\alias{mcnode}
\alias{mcdata}
\alias{mcdatanocontrol}
\title{Build mcnode Objects from Data}
\usage{
mcdata(
  data,
  type = c("V", "U", "VU", "0"),
  nsv = ndvar(),
  nsu = ndunc(),
  nvariates = 1,
  outm = "each"
)

mcdatanocontrol(
  data,
  type = c("V", "U", "VU", "0"),
  nsv = ndvar(),
  nsu = ndunc(),
  nvariates = 1,
  outm = "each"
)
}
\arguments{
\item{data}{the numeric/logical vector/matrix/array of data or a \samp{mcnode} object.}

\item{type}{the type of node to be built. By default, a \samp{"V"} node.}

\item{nsv}{the variability dimension (\samp{type="V"} or \samp{type="VU"}) of the node.
By default: the current value in \code{\link{mc.control}}.}

\item{nsu}{the uncertainty dimension (\samp{type="U"} or \samp{type="VU"}) of the node.
By default: the current value in \code{\link{mc.control}}.}

\item{nvariates}{the number of variates. By default: 1.}

\item{outm}{the output of the \samp{mcnode} for multivariate nodes. May be \samp{"each"}
(default), \samp{"none"}, or a vector of function names applied on the variates dimension
before any output.}
}
\value{
An \samp{mcnode} object.
}
\description{
Creates a \samp{mcnode} object from a vector, an array or another \samp{mcnode}.

\samp{mcdatanocontrol} is a fast but unchecked version of \code{\link{mcdata}} which
forces the dimension of data to be \samp{(nsv x nsu x nvariates)} and sets the attributes
and class without any validation. Useful when your model is tested and speed is important.
}
\details{
A \samp{mcnode} object is the basic element of a \code{\link{mc}} object.
It is an array of dimension \samp{(nsv x nsu x nvariates)}. Four types exist:
\describe{
  \item{\samp{"V" mcnode}}{for "Variability", arrays of dimension \samp{(nsv x 1 x nvariates)}.
    The variability of the data should reflect parameter variability.}
  \item{\samp{"U" mcnode}}{for "Uncertainty", arrays of dimension \samp{(1 x nsu x nvariates)}.
    The variability reflects parameter uncertainty.}
  \item{\samp{"VU" mcnode}}{for "Variability and Uncertainty", arrays of dimension
    \samp{(nsv x nsu x nvariates)}.}
  \item{\samp{"0" mcnode}}{for "Neither Variability nor Uncertainty", arrays of dimension
    \samp{(1 x 1 x nvariates)}.}
}

Multivariate nodes (\samp{nvariates != 1}) should be used for multivariate distributions
(\code{\link{rmultinomial}}, \code{\link{rmultinormal}}, \code{\link{rempiricalD}},
\code{\link{rdirichlet}}).

Recycling rules are limited: recycling is only permitted to fill a dimension from 1 to
the final size of the dimension.
}
\examples{
oldvar <- ndvar()
oldunc <- ndunc()
ndvar(3)
ndunc(5)

(x0 <- mcdata(100, type = "0"))
(xV <- mcdata(1:ndvar(), type = "V"))
(xU <- mcdata(10 * 1:ndunc(), type = "U"))
(xVU <- mcdata(1:(ndvar() * ndunc()), type = "VU"))

## Multivariates
(x0M <- mcdata(1:2, type = "0", nvariates = 2))
(xVM <- mcdata(1:(2 * ndvar()), type = "V", nvariates = 2))

ndvar(oldvar)
ndunc(oldunc)
}
\seealso{
\code{\link{mcstoc}} to build a stochastic \samp{mcnode} object.
  \code{\link{mc}} to build a Monte-Carlo object.
  \code{\link{is.mcnode}}, \code{\link{dimmcnode}}, \code{\link{typemcnode}}.
}
\keyword{methods}
