% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcprobtree.R
\name{mcprobtree}
\alias{mcprobtree}
\title{Creates a Stochastic mcnode Object Using a Probability Tree}
\usage{
mcprobtree(
  mcswitch,
  mcvalues,
  type = c("V", "U", "VU", "0"),
  nsv = ndvar(),
  nsu = ndunc(),
  nvariates = 1,
  outm = "each",
  seed = NULL
)
}
\arguments{
\item{mcswitch}{a vector of probabilities/weights or an \samp{mcnode}.}

\item{mcvalues}{a named list of \samp{mcnode}s, \samp{mcdata} calls, or \samp{mcstoc} calls.
Each element should be or lead to a compatible \samp{mcnode}.}

\item{type}{the type of \samp{mcnode} to be built. By default, a \samp{"V"} node.}

\item{nsv}{the number of simulations in the variability dimension.}

\item{nsu}{the number of simulations in the uncertainty dimension.}

\item{nvariates}{the number of variates of the final \samp{mcnode}.}

\item{outm}{the default output for multivariate nodes. See \code{\link{outm}}.}

\item{seed}{the random seed. If \samp{NULL} the seed is unchanged.}
}
\value{
An \samp{mcnode} object.
}
\description{
Builds an \samp{mcnode} as a mixture of \samp{mcnode} objects.
}
\details{
\samp{mcswitch} may be either:
\itemize{
  \item a vector of weights (need not sum to one, must be nonneg and not all zero).
    Length must equal the number of elements in \samp{mcvalues}.
  \item a \samp{"0 mcnode"} to build any type of node.
  \item a \samp{"V mcnode"} to build a \samp{"V"} or \samp{"VU"} node.
  \item a \samp{"U mcnode"} to build a \samp{"U"} or \samp{"VU"} node.
  \item a \samp{"VU mcnode"} to build a \samp{"VU"} node.
}

Names of elements in \samp{mcvalues} should correspond to values in \samp{mcswitch},
specified as character. Elements are evaluated only if needed.
}
\examples{
conc1 <- mcstoc(rnorm, type="VU", mean=10, sd=2)
conc2 <- mcstoc(runif, type="VU", min=-6, max=-5)
conc3 <- mcdata(0, type="VU")
## Randomly in the cells
whichdist <- mcstoc(rempiricalD, type="VU", values=1:3, prob=c(.75, .20, .05))
mcprobtree(whichdist, list("1"=conc1, "2"=conc2, "3"=conc3), type="VU")
## Equivalent using weights
mcprobtree(c(.75, .20, .05), list("1"=conc1, "2"=conc2, "3"=conc3), type="VU")
}
\seealso{
\code{\link{mcdata}}, \code{\link{mcstoc}}, \code{\link{switch}}.
}
\keyword{methods}
