% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcratio.R
\name{mcratio}
\alias{mcratio}
\title{Ratio of Uncertainty to Variability}
\usage{
mcratio(x, pcentral = 0.5, pvar = 0.975, punc = 0.975, na.rm = FALSE)
}
\arguments{
\item{x}{an \samp{mc} or an \samp{mcnode} object.}

\item{pcentral}{the quantile for the central tendency.}

\item{pvar}{the quantile for the measure of variability.}

\item{punc}{the quantile for the measure of uncertainty.}

\item{na.rm}{logical; should NA values be stripped before computation?}
}
\value{
A matrix.
}
\description{
Provides measures of variability, uncertainty, and both combined for an
\samp{mc} or an \samp{mcnode} object.
}
\details{
Given quantiles A, B, C, D defined as:
\describe{
  \item{A}{the \samp{(100*pcentral)}th percentile of uncertainty for the \samp{(100*pcentral)}th percentile of variability}
  \item{B}{the \samp{(100*pcentral)}th percentile of uncertainty for the \samp{(100*pvar)}th percentile of variability}
  \item{C}{the \samp{(100*punc)}th percentile of uncertainty for the \samp{(100*pcentral)}th percentile of variability}
  \item{D}{the \samp{(100*punc)}th percentile of uncertainty for the \samp{(100*pvar)}th percentile of variability}
}
the following ratios are estimated: Variability Ratio: B/A; Uncertainty Ratio: C/A;
Overall Uncertainty Ratio: D/A.
}
\examples{
data(total)
mcratio(total, na.rm=TRUE)
}
\references{
Ozkaynak, H., Frey, H.C., Burke, J., Pinder, R.W. (2009)
  "Analysis of coupled model uncertainties in source-to-dose modeling of human
  exposures to ambient air pollution: A PM2.5 case study", Atmospheric Environment,
  43(9), 1641-1649.
}
\keyword{distribution}
