% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mcstoc.R
\name{mcstoc}
\alias{mcstoc}
\title{Creates Stochastic mcnode Objects}
\usage{
mcstoc(
  func = runif,
  type = c("V", "U", "VU", "0"),
  ...,
  nsv = ndvar(),
  nsu = ndunc(),
  nvariates = 1,
  outm = "each",
  nsample = "n",
  seed = NULL,
  rtrunc = FALSE,
  linf = -Inf,
  lsup = Inf,
  lhs = FALSE
)
}
\arguments{
\item{func}{a function providing random data or its name as character.}

\item{type}{the type of \samp{mcnode} to be built. By default, a \samp{"V"} node.
See \code{\link{mcnode}} for details.}

\item{...}{all other arguments but the size of the sample to be passed to \samp{func}.
These arguments should be vectors or \samp{mcnode}s (arrays prohibited).}

\item{nsv}{the number of simulations in the variability dimension.}

\item{nsu}{the number of simulations in the uncertainty dimension.}

\item{nvariates}{the number of variates of the output.}

\item{outm}{the output of the \samp{mcnode} for multivariate nodes. May be \samp{"each"}
(default), \samp{"none"}, or a vector of function names (as character strings) applied
on the variates dimension before output (e.g. \samp{"mean"}, \samp{c("min","max")}).}

\item{nsample}{the name of the parameter of the function giving the size of the vector.
By default \samp{"n"}, as in most distributions of the \pkg{stats} library.}

\item{seed}{the random seed used for the evaluation. If \samp{NULL} the seed is unchanged.}

\item{rtrunc}{should the distribution be truncated? See \code{\link{rtrunc}}.}

\item{linf}{if truncated: lower limit. May be a scalar, an array or a \samp{mcnode}.}

\item{lsup}{if truncated: upper limit. May be a scalar, an array or a \samp{mcnode}.
\samp{lsup} must be pairwise strictly greater than \samp{linf}.}

\item{lhs}{should a Latin Hypercube Sampling be used? See \code{\link{lhs}}.}
}
\value{
An \samp{mcnode} object.
}
\description{
Creates a \code{\link{mcnode}} object using a random generating function.
}
\details{
Any function that accepts vectors/matrices as arguments may be used (notably all current
random generators of the \pkg{stats} package). Arguments may be sent classically, but it
is strongly recommended to use consistent \samp{mcnode}s if arguments should be recycled,
since complex recycling is handled for \samp{mcnode} but not for vectors.

Compatibility rules for \samp{mcnode} arguments:
\describe{
  \item{\samp{type="V"}}{accepts \samp{"0" mcnode} of dim \samp{(1 x 1 x nvariates)} or
    \samp{(1 x 1 x 1)} (recycled) and \samp{"V" mcnode} of dim \samp{(nsv x 1 x nvariates)}
    or \samp{(nsv x 1 x 1)} (recycled).}
  \item{\samp{type="U"}}{accepts \samp{"0" mcnode} of dim \samp{(1 x 1 x nvariates)} or
    \samp{(1 x 1 x 1)} (recycled) and \samp{"U" mcnode} of dim \samp{(1 x nsu x nvariates)}
    or \samp{(1 x nsu x 1)} (recycled).}
  \item{\samp{type="VU"}}{accepts \samp{"0"}, \samp{"V"}, \samp{"U"} and \samp{"VU"} nodes
    with appropriate recycling rules.}
  \item{\samp{type="0"}}{accepts \samp{"0" mcnode} of dim \samp{(1 x 1 x nvariates)} or
    \samp{(1 x 1 x 1)} (recycled).}
}

If \samp{rtrunc=TRUE}, the distribution is truncated on \samp{(linf, lsup]}. The function
\samp{func} should have a \samp{q} form and a \samp{p} form.

If \samp{lhs=TRUE}, a Latin Hypercube Sampling is used on \samp{nsv} and \samp{nsu}.
The function \samp{func} should have a \samp{q} form. Not allowed with multivariate
distributions.
}
\examples{
Oldnvar <- ndvar()
Oldnunc <- ndunc()
ndvar(5)
ndunc(4)

## Compatibility with mcdata as arguments
x0 <- mcstoc(runif, type = "0")
xV <- mcstoc(runif, type = "V")
xU <- mcstoc(runif, type = "U")
xVU <- mcstoc(runif, type = "VU")

## "V" accepts "0" and "V" mcdata
mcstoc(rnorm, type = "V", mean = x0, sd = xV)

## "VU" accepts "0", "U", "V" and "VU" with correct recycling
mcstoc(rnorm, type = "VU", mean = xV, sd = xU)

ndvar(Oldnvar)
ndunc(Oldnunc)
}
\seealso{
\code{\link{mcnode}} for a description of \samp{mcnode} objects.
  \code{\link{Ops.mcnode}} for operations on \samp{mcnode} objects.
  \code{\link{rtrunc}} for important warnings on the \samp{rtrunc} option.
}
\keyword{methods}
