% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmultinormal.R
\name{multinormal}
\alias{multinormal}
\alias{rmultinormal}
\alias{dmultinormal}
\title{The Vectorized Multivariate Normal Distribution}
\usage{
rmultinormal(n, mean, sigma, method = c("eigen", "svd", "chol"))

dmultinormal(x, mean, sigma, log = FALSE)
}
\arguments{
\item{n}{number of observations. If \samp{length(n) > 1}, the length is taken to be the
number required.}

\item{mean}{vector or matrix of means. If a matrix, each row is taken to be a mean vector.
Default is a vector of 0 of convenient length.}

\item{sigma}{covariance vector corresponding to the coercion of the covariance matrix into
a vector (if unique for all \samp{n} or \samp{x}) or array of covariance vectors (if
varying). Default is a diagonal matrix of convenient size.}

\item{method}{matrix decomposition used to determine the matrix root of sigma. Possible
methods are eigenvalue decomposition (\samp{"eigen"}, default), singular value
decomposition (\samp{"svd"}), and Cholesky decomposition (\samp{"chol"}).}

\item{x}{vector or matrix of quantiles. If x is a matrix, each row is taken to be a quantile.}

\item{log}{logical; if \samp{TRUE}, densities d are given as \samp{log(d)}.}
}
\description{
Vectorized versions of \code{\link[mvtnorm]{rmvnorm}} and \code{\link[mvtnorm]{dmvnorm}}
from the \pkg{mvtnorm} package, providing random deviates and density for the multivariate
normal distribution with varying vectors of means and varying covariance matrices.
}
\details{
\samp{rmvnorm(n, m, s)} is equivalent to \samp{rmultinormal(n, m, as.vector(s))}.
\samp{dmvnorm(x, m, s)} is equivalent to \samp{dmultinormal(x, m, as.vector(s))}.

If \samp{mean} and/or \samp{sigma} is a matrix, the first random deviate will use
the first row of \samp{mean} and/or \samp{sigma}, the second will use the second row, ...
recycling being permitted by row.

If \samp{mean} is a vector of length \samp{l} or is a matrix with \samp{l} columns,
\samp{sigma} should be a vector of length \samp{l x l} or a matrix with \samp{l^2} columns.
}
\note{
The use of a varying sigma may be very time consuming.
}
\examples{
## mean and sigma as vectors
(mean <- c(10, 0))
(sigma <- matrix(c(1, 2, 2, 10), ncol = 2))
sigma <- as.vector(sigma)
(x <- matrix(c(9, 8, 1, -1), ncol = 2))
round(rmultinormal(10, mean, sigma))
dmultinormal(x, mean, sigma)

## mean as matrix
(mean <- matrix(c(10, 0, 0, 10), ncol = 2))
round(rmultinormal(10, mean, sigma))
}
\keyword{distribution}
