% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is.na.R
\name{NA.mcnode}
\alias{NA.mcnode}
\alias{is.na.mcnode}
\alias{is.nan.mcnode}
\alias{is.finite.mcnode}
\alias{is.infinite.mcnode}
\title{Finite, Infinite, NA and NaN Numbers in mcnode}
\usage{
\method{is.na}{mcnode}(x)

\method{is.nan}{mcnode}(x)

\method{is.finite}{mcnode}(x)

\method{is.infinite}{mcnode}(x)
}
\arguments{
\item{x}{a \samp{mcnode} object.}
}
\value{
a logical \samp{mcnode} object.
}
\description{
\samp{is.na}, \samp{is.nan}, \samp{is.finite} and \samp{is.infinite} return a logical
\samp{mcnode} of the same dimension as \samp{x}.
}
\examples{
x <- log(mcstoc(rnorm, nsv=1001))
x
is.na(x)
}
\seealso{
\code{\link{is.finite}}, \code{\link{NA}}
}
\keyword{NA}
