% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ops.mcnode.R
\name{Ops.mcnode}
\alias{Ops.mcnode}
\title{Operations on mcnode Objects}
\usage{
\method{Ops}{mcnode}(e1, e2)
}
\arguments{
\item{e1}{an \samp{mcnode} object, a vector or an array.}

\item{e2}{an optional \samp{mcnode} object, a vector or an array.}
}
\value{
the result as a \samp{mcnode} object.
}
\description{
Alters the way operations are performed on \samp{mcnode} objects for better consistency.
This method is used for any of the Group \code{\link{Ops}} functions.
}
\details{
Rules (illustrated with \samp{+}, ignoring the \samp{nvariates} dimension):
\itemize{
  \item \samp{0 + 0 = 0}; \samp{0 + V = V}; \samp{0 + U = U}; \samp{0 + VU = VU}
  \item \samp{V + V = V}: if both of the same \samp{(nsv)} dimension
  \item \samp{V + U = VU}: U recycled "by row", V recycled "by column"
  \item \samp{V + VU = VU}: V recycled "by column"
  \item \samp{U + U = U}: if both of the same \samp{(nsu)} dimension
  \item \samp{U + VU = VU}: U recycled "by row"
  \item \samp{VU + VU = VU}: if dimensions are \samp{(nsu x nsv)}
}

The \samp{outm} attribute is transferred as: \samp{each + each = each};
\samp{none + other = other}; \samp{other1 + other2 = other1}.
}
\examples{
oldvar <- ndvar()
oldunc <- ndunc()
ndvar(30)
ndunc(20)
x0 <- mcdata(3, type="0")
xV <- mcdata(1:ndvar(), type="V")
xU <- mcdata(1:ndunc(), type="U")
xVU <- mcdata(1:(ndunc()*ndvar()), type="VU")
-x0
x0 + 3
xV * x0
xV + xU
xU + xVU
ndvar(oldvar)
ndunc(oldunc)
}
\seealso{
\code{\link{mcdata}}, \code{\link{mcstoc}}
}
\keyword{utilities}
