% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outm.R
\name{outm}
\alias{outm}
\title{Changes the Output of Nodes}
\usage{
outm(x, value = "each", which.node = 1)
}
\arguments{
\item{x}{a \samp{mcnode} or a \samp{mc} object.}

\item{value}{the output of the \samp{mcnode} for multivariate nodes. May be \samp{"each"}
(default) if output should be provided for each variate independently, \samp{"none"} for
no output, or a vector of function names (as character strings) applied on the variates
dimension before any output (e.g. \samp{"mean"}, \samp{"median"}, \samp{c("min","max")}).
The function should return one value per vector.}

\item{which.node}{which node should be changed in a \samp{mc} object.}
}
\value{
\samp{x} with a modified \samp{outm} attribute.
}
\description{
Changes the \samp{outm} attribute of an \samp{mcnode} or a node of an \samp{mc} object.
}
\examples{
data(total)
total$xVUM2
## since outm = NULL
summary(total$xVUM2)
x <- outm(total$xVUM2, c("min"))
summary(x)
}
\keyword{misc}
