% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.mc.R
\name{plot.mc}
\alias{plot.mc}
\alias{plot.mcnode}
\alias{plot.plotmc}
\title{Plots Results of a Monte Carlo Simulation}
\usage{
\method{plot}{mc}(
  x,
  prec = 0.001,
  stat = c("median", "mean"),
  lim = c(0.025, 0.25, 0.75, 0.975),
  na.rm = TRUE,
  griddim = NULL,
  xlab = NULL,
  ylab = "Fn(x)",
  main = "",
  draw = TRUE,
  paint = TRUE,
  xlim = NULL,
  ylim = NULL,
  ...
)

\method{plot}{mcnode}(x, ...)

\method{plot}{plotmc}(x, ...)
}
\arguments{
\item{x}{a \samp{mcnode} or a \samp{mc} object.}

\item{prec}{the precision of the plot. \samp{0.001} gives an ecdf from
the 0.000, 0.001, ..., 1.000 quantiles.}

\item{stat}{the function used for estimates (2D \samp{mc} or \samp{mcnode}). By default the median.}

\item{lim}{a vector of numbers (between 0 and 1) indicating the envelope. May be \samp{NULL} or empty.}

\item{na.rm}{should NA values be discarded?}

\item{griddim}{a vector of two integers for the grid size. If \samp{NULL}, calculated automatically.}

\item{xlab}{vector of labels for the x-axis. If \samp{NULL}, use the name of the node.}

\item{ylab}{vector of labels for the y-axis.}

\item{main}{vector of main titles of the graph.}

\item{draw}{should the plot be drawn?}

\item{paint}{should the envelopes be filled?}

\item{xlim}{x coordinate range: a vector of length 2 or a list of such vectors.}

\item{ylim}{y coordinate range: a vector of length 2 or a list of such vectors.}

\item{...}{further arguments to be passed to \samp{plot.stepfun}.}
}
\value{
a \samp{plot.mc} object (list of quantiles used to draw the plot), invisibly.
}
\description{
Plots the empirical cumulative distribution function of a \samp{mcnode} or a \samp{mc}
object (\samp{"0"} and \samp{"V"} nodes) or the empirical CDF of the estimate of a
\samp{mcnode} or \samp{mc} object (\samp{"U"} and \samp{"VU"} nodes).
}
\details{
For \samp{"VU"} and \samp{"U"} nodes, quantiles are calculated using
\code{\link{quantile.mc}} within each of the \samp{nsu} simulations (by columns). The
medians (or means if \samp{stat="mean"}) are plotted, with \samp{lim} quantiles as
the envelope.
}
\examples{
data(total)
plot(xVUM3)
## Only one envelope for 0.025 and 0.975
plot(xVUM3, lim=c(0.025, 0.975))
}
\references{
Cullen AC and Frey HC (1999) Probabilistic techniques in exposure assessment.
  Plenum Press, USA, pp. 81-155.
}
\seealso{
\code{\link{ecdf}}, \code{\link{plot}}, \code{\link{quantile.mc}},
  \code{\link{ggplotmc}} for a ggplot2 version.
}
\keyword{hplot}
