% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.tornado.R
\name{plot.tornado}
\alias{plot.tornado}
\alias{plot.tornadounc}
\title{Draw a Tornado Chart}
\usage{
\method{plot}{tornado}(
  x,
  which = 1,
  name = NULL,
  stat = c("median", "mean"),
  xlab = "method",
  ylab = "",
  ...
)

\method{plot}{tornadounc}(
  x,
  which = 1,
  stat = "median",
  name = NULL,
  xlab = "method",
  ylab = "",
  ...
)
}
\arguments{
\item{x}{a \code{\link{tornado}} or \code{\link{tornadounc}} object.}

\item{which}{which output to plot (for multivariate output).}

\item{name}{vector of names for input variables. If \samp{NULL}, taken from the object.}

\item{stat}{the statistic of the output to be considered. For a \samp{tornado} object:
\samp{"median"} or \samp{"mean"}. For a \samp{tornadounc} object: should match a row
name, or use the row number.}

\item{xlab}{label of the x axis. If \samp{"method"}, use the correlation method.}

\item{ylab}{label of the y axis.}

\item{...}{further arguments to be passed to the \samp{plot} function.}
}
\value{
\samp{NULL} (called for side effects).
}
\description{
Draws a Tornado chart as provided by \code{\link{tornado}} or \code{\link{tornadounc}}.
}
\details{
A point is drawn at the estimate, and the segment reflects the uncertainty around it.
}
\examples{
data(ec)
x <- evalmcmod(ec$modEC2, nsv=100, nsu=100, seed=666)
tor <- tornado(x, 7)
plot(tor)
}
\seealso{
\code{\link{tornado}}
}
\keyword{hplot}
