% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pmin.R
\name{pmin.mcnode}
\alias{pmin.mcnode}
\alias{pmin}
\alias{pmax}
\alias{pmin.default}
\alias{pmax.default}
\alias{pmax.mcnode}
\title{Parallel Maxima and Minima for mcnodes}
\usage{
pmin(..., na.rm = FALSE)

pmax(..., na.rm = FALSE)

\method{pmin}{default}(..., na.rm = FALSE)

\method{pmax}{default}(..., na.rm = FALSE)

\method{pmin}{mcnode}(..., na.rm = FALSE)

\method{pmax}{mcnode}(..., na.rm = FALSE)
}
\arguments{
\item{...}{one or more \samp{mcnode}s or \samp{mcnode}s and compatible vectors.
The first element must be an \samp{mcnode}.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
an \samp{mcnode} of adequate type and dimension.
}
\description{
Returns the parallel maxima and minima of the input values, extended to \samp{mcnode} objects.
The same rules as in \code{\link{Ops.mcnode}} are applied for type determination.
}
\details{
\samp{pmax} and \samp{pmin} take one or more \samp{mcnode} and/or vectors as arguments and
return a \samp{mcnode} of adequate type and size giving the "parallel" maxima or minima.
Note that the first element of \samp{...} should be an \samp{mcnode}.
}
\examples{
ndvar(10)
ndunc(21)
x <- mcstoc(rnorm, "V")
pmin(x, 0)
y <- mcdata(rep(c(-1, 1), length=ndunc()), "U")
unclass(pmin(x, y))
}
\seealso{
\code{\link{min}}, \code{\link{Ops.mcnode}}
}
\keyword{utilities}
