% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mc.R
\name{print.mc}
\alias{print.mc}
\alias{print.mcnode}
\title{Print a mcnode or mc Object}
\usage{
\method{print}{mc}(x, digits = 3, ...)

\method{print}{mcnode}(x, ...)
}
\arguments{
\item{x}{a \samp{mcnode} or a \samp{mc} object.}

\item{digits}{number of digits to be used.}

\item{...}{further arguments to be passed to the print function.}
}
\value{
an invisible data frame.
}
\description{
Prints a description of the structure of a \samp{mc} or \samp{mcnode} object.
}
\examples{
data(total)
print(xVU)
print(total)
}
\seealso{
\code{\link{mcnode}} for \samp{mcnode} objects, \code{\link{mc}} for \samp{mc} objects.
}
\keyword{print}
