% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quantile.mc.R
\name{quantile.mc}
\alias{quantile.mc}
\alias{quantile.mcnode}
\title{Quantiles of a mc or mcnode Object}
\usage{
\method{quantile}{mc}(x, probs = seq(0, 1, 0.01), lim = c(0.025, 0.975), na.rm = TRUE, ...)

\method{quantile}{mcnode}(x, ...)
}
\arguments{
\item{x}{a \samp{mc} or \samp{mcnode} object.}

\item{probs}{the quantiles to be calculated.}

\item{lim}{a vector of numbers (between 0 and 1) indicating the envelope. May be \samp{NULL} or empty.}

\item{na.rm}{logical; should NA values be stripped?}

\item{...}{for generic method consistency.}
}
\value{
a list of quantiles.
}
\description{
Evaluates quantiles of a \samp{mc} or \samp{mcnode} object. Used internally by
\code{\link{plot.mc}}.
}
\details{
Quantiles are evaluated in the variability dimension. Then, the median, the mean and
the \samp{lim} quantiles are evaluated for each of these quantiles in the uncertainty dimension.
}
\examples{
data(total)
quantile(total$xVUM3)
quantile(total)
}
\seealso{
\code{\link{plot.mc}}, \code{\link{quantile}}.
}
\keyword{univar}
