% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rtrunc.R
\name{rtrunc}
\alias{rtrunc}
\title{Random Truncated Distributions}
\usage{
rtrunc(distr = runif, n, linf = -Inf, lsup = Inf, ...)
}
\arguments{
\item{distr}{A function providing random data or its name as character.
The function \samp{rdistr} should have a \samp{qdistr} form (with argument \samp{p})
and a \samp{pdistr} form (with argument \samp{q}). Example: \samp{rnorm}
(has a \samp{qnorm} and a \samp{pnorm} form), \samp{rbeta}, \samp{rbinom},
\samp{rgamma}, ...}

\item{n}{the size of the sample.}

\item{linf}{a vector of lower bounds.}

\item{lsup}{a vector of upper bounds, with \samp{lsup > linf} (strictly).}

\item{...}{all arguments to be passed to \samp{pdistr} and \samp{qdistr}.}
}
\value{
A vector of \samp{n} values.
}
\description{
Provides samples from classical \R distributions and \samp{mc2d} specific
distributions truncated between \samp{linf} (excluded) and \samp{lsup} (included).
}
\details{
The function 1) evaluates the \samp{p} values corresponding to \samp{linf} and
\samp{lsup} using \samp{pdistr}; 2) samples \samp{n} values using
\samp{runif(n, min=pinf, max=psup)}, and 3) takes the \samp{n} corresponding
quantiles from the specified distribution using \samp{qdistr}.

All distributions (but \code{sample}) implemented in the \pkg{stats} library
could be used. The arguments in \dots should be named. Do not use \samp{log},
\samp{log.p} or \samp{lower.tail}.

For discrete distributions, \samp{rtrunc} samples within \samp{(linf, lsup]}.

\strong{Warning:} The method is flexible, but can lead to problems linked to rounding
errors in some extreme situations. The function checks that the values are in the
expected range and returns an error if not.
}
\examples{
rtrunc("rnorm", n = 10, linf = 0)
range(rtrunc(rnorm, n = 1000, linf = 3, lsup = 5, sd = 10))
## Discrete distributions
range(rtrunc(rpois, 1000, linf = 2, lsup = 4, lambda = 1))
}
\keyword{distribution}
