% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.mc.R, R/summary.mccut.R
\name{summary.mc}
\alias{summary.mc}
\alias{summary.mcnode}
\alias{print.summary.mc}
\alias{summary.mccut}
\title{Summary of mcnode and mc Object}
\usage{
\method{summary}{mc}(
  object,
  probs = c(0, 0.025, 0.25, 0.5, 0.75, 0.975, 1),
  lim = c(0.025, 0.975),
  ...
)

\method{summary}{mcnode}(
  object,
  probs = c(0, 0.025, 0.25, 0.5, 0.75, 0.975, 1),
  lim = c(0.025, 0.975),
  digits = 3,
  ...
)

\method{print}{summary.mc}(x, digits = 3, ...)

\method{summary}{mccut}(object, lim = c(0.025, 0.975), ...)
}
\arguments{
\item{object}{a \samp{mcnode}, \samp{mc} or \samp{mccut} object.}

\item{probs}{a vector of values used for the quantile function (variability dimension).}

\item{lim}{a vector of values used for the quantile function (uncertainty dimension).}

\item{...}{for generic function consistency.}

\item{digits}{number of digits in the print.}

\item{x}{a \samp{summary.mc} object as returned by the \samp{summary.mc} function.}
}
\value{
a list.
}
\description{
Provides a summary of a \samp{mcnode}, a \samp{mc} or a \samp{mccut} object.
}
\details{
The mean, the standard deviation and the \samp{probs} quantiles will be evaluated in the
variability dimension. The median, the mean and the \samp{lim} quantiles will then be
evaluated on these statistics in the uncertainty dimension.

For multivariate nodes: if the \samp{"outm"} attribute is \samp{"none"}, the node is not
evaluated; if it is \samp{"each"}, the variates are evaluated one by one; if it is a function
name (e.g. \samp{"mean"}), that function is applied on the variates dimension before output.
}
\examples{
data(total)
summary(xVUM3)
summary(total)
}
\seealso{
\code{\link{mcnode}} for mcnode objects, \code{\link{mc}} for mc objects,
  \code{\link{mccut}} for mccut objects, \code{\link{quantile}}.
}
\keyword{univar}
