% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornado.R
\name{tornado}
\alias{tornado}
\alias{print.tornado}
\title{Computes Correlation Between Inputs and Output in the Variability Dimension (Tornado)}
\usage{
tornado(
  mc,
  output = length(mc),
  use = "all.obs",
  method = c("spearman", "kendall", "pearson"),
  lim = c(0.025, 0.975)
)

\method{print}{tornado}(x, ...)
}
\arguments{
\item{mc}{a \code{\link{mc}} or \code{\link{mccut}} object.}

\item{output}{the rank or name of the output to be considered. By default: the last element.}

\item{use}{an optional character string for computing covariances in the presence of
missing values: \samp{"all.obs"}, \samp{"complete.obs"} or \samp{"pairwise.complete.obs"}.}

\item{method}{a character string for the correlation coefficient: \samp{"spearman"}
(default), \samp{"kendall"} or \samp{"pearson"}.}

\item{lim}{a vector of quantiles used to compute the credible interval.}

\item{x}{a \samp{tornado} object.}

\item{...}{further arguments to be passed to the final print function.}
}
\value{
an invisible object of class \samp{tornado}, containing: \samp{value} (correlation
  coefficients), \samp{output} (name of output), \samp{method}, \samp{use}.
}
\description{
Provides statistics for a tornado chart. Evaluates correlations between output and inputs
of a \samp{mc} object in the variability dimension.
}
\details{
The tornado function computes Spearman's rho statistic to estimate a rank-based measure
of association between one set of random variables of a \samp{mc} object (the output)
and the others (the inputs).

\samp{tornado} may be applied on a \samp{mccut} object if a \samp{tornado} function was
used in the third block of the \code{\link{evalmccut}} call.

Type rules for the output node:
\itemize{
  \item \samp{"V" mcnode}: correlations are only provided for other \samp{"V" mcnode}s.
  \item \samp{"VU" mcnode}: correlations are provided for other \samp{"VU"} and \samp{"V"} nodes.
}
}
\examples{
data(total)
tornado(total, 2, "complete.obs", "spearman", c(0.025, 0.975))
(y <- tornado(total, 10, "complete.obs", "spearman", c(0.025, 0.975)))
plot(y)
}
\seealso{
\code{\link{cor}}, \code{\link{plot.tornado}} to draw the results.
}
\keyword{univar}
