% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tornadounc.mc.R, R/tornadounc.mccut.R
\name{tornadounc}
\alias{tornadounc}
\alias{tornadounc.default}
\alias{tornadounc.mc}
\alias{print.tornadounc}
\alias{tornadounc.mccut}
\title{Computes Correlation Between Inputs and Output in the Uncertainty Dimension (Tornado)}
\usage{
tornadounc(mc, ...)

\method{tornadounc}{default}(mc, ...)

\method{tornadounc}{mc}(
  mc,
  output = length(mc),
  quant = c(0.5, 0.75, 0.975),
  use = "all.obs",
  method = c("spearman", "kendall", "pearson"),
  ...
)

\method{print}{tornadounc}(x, ...)

\method{tornadounc}{mccut}(
  mc,
  output = length(mc),
  quant = c(0.5, 0.75, 0.975),
  use = "all.obs",
  method = c("spearman", "kendall", "pearson"),
  ...
)
}
\arguments{
\item{mc}{a \samp{mc} object.}

\item{...}{further arguments to be passed to the final print function.}

\item{output}{the rank or name of the output. Should be a \samp{"VU"} or \samp{"U"} node.
By default: the last element.}

\item{quant}{the vector of quantiles used in the variability dimension.}

\item{use}{an optional character string for computing covariances in presence of missing
values: \samp{"all.obs"}, \samp{"complete.obs"} or \samp{"pairwise.complete.obs"}.}

\item{method}{a character string for the correlation coefficient: \samp{"spearman"}
(default), \samp{"kendall"} or \samp{"pearson"}.}

\item{x}{a \samp{tornadounc} object.}
}
\value{
an invisible object of class \samp{tornadounc}.
}
\description{
Provides statistics for a tornado chart. Evaluates correlations between output and inputs
of a \samp{mc} object in the uncertainty dimension.
}
\details{
The function computes Spearman's rho statistic between values or statistics calculated in
the variability dimension of inputs and outputs. The statistics are the mean, the standard
deviation and the quantiles specified by \samp{quant}.

\samp{tornadounc.mccut} may be applied on a \code{\link{mccut}} object if a
\samp{summary.mc} function was used in the third block of the \code{\link{evalmccut}} call.
}
\examples{
data(total)
tornadounc(total, 3)
(y <- tornadounc(total, 10, use="complete.obs"))
plot(y, 1, 1)
}
\seealso{
\code{\link{cor}}, \code{\link{tornado}} for variability dimension,
  \code{\link{plot.tornadounc}} to draw the results.
}
\keyword{univar}
