% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unmc.R
\name{unmc}
\alias{unmc}
\title{Unclass the mc or mcnode Object}
\usage{
unmc(x, drop = TRUE)
}
\arguments{
\item{x}{a \samp{mc} or a \samp{mcnode} object.}

\item{drop}{should dimensions of size 1 be dropped (see \code{\link{drop}}).}
}
\value{
If \samp{x} is an \samp{mc} object: a list of arrays. If \samp{drop=TRUE},
  a list of vectors, matrices and arrays.
  If \samp{x} is an \samp{mcnode} object: an array. If \samp{drop=TRUE},
  a vector, matrix or array.
}
\description{
Unclasses an \samp{mc} object into a list of arrays, or an \samp{mcnode} object into
an array.
}
\examples{
data(total)
## A vector
unmc(total$xV, drop=TRUE)
## An array
unmc(total$xV, drop=FALSE)
}
\keyword{manip}
