\name{print.mc}
\alias{print.mc}
\alias{print.mcnode}
\title{
Prints a mcnode or a mc Object
}
\description{
Print a description of the structure of the mc or the mcnode object.
}
\usage{
\method{print}{mc}(x, digits=3, ...)
\method{print}{mcnode}(x, ...)
}
\arguments{
\item{x}{a mcnode or a mc object.}
\item{digits}{Number of digits to be used.}
\item{\dots}{Further arguments to be passed to the print function.}
}
\value{
An invisible data frame.
}
\author{
Regis Pouillot
}
\seealso{
\code{\link{mcnode}} for mcnode objects. \code{\link{mc}} for mc
objects.
}
\keyword{print}
