\name{print.mc}
\alias{print.mc}
\alias{print.mcnode}
\title{
Prints a mcnode or a mc Object
}
\description{
Print a description of the structure of the \samp{mc} or the
\samp{mcnode} object.
}
\usage{
\method{print}{mc}(x, digits=3, ...)
\method{print}{mcnode}(x, ...)
}
\arguments{
\item{x}{a \samp{mcnode} or a \samp{mc} object.}
\item{digits}{Number of digits to be used.}
\item{\dots}{Further arguments to be passed to the print function.}
}
\value{
An invisible data frame.
}
\seealso{
\code{\link{mcnode}} for \samp{mcnode} objects. \code{\link{mc}} for
\samp{mc} objects.
}
\keyword{print}
